requirements.txt
pyyaml
typer
feedparser
pytest
requests
